♻️ Migrate all the DTO models into the fitting module.
